services/web-app/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web"
version = "0.1.0"
requires-python=">=3.12"
dependencies = [
    "gradio==5.33.0",
    "hydra-core==1.3.2",
    "requests==2.32.3"
]

[project.optional-dependencies]
dev = [
    "uvicorn==0.34.3",
    "fastapi==0.115.12",
    "pylint>=4.0.4",
    "autopep8>=2.3.2",
    "mypy>=1.19.0",
    "pre-commit>=4.5.0"
]
